requirements.txt
google-cloud-pubsub==1.3.1
requests==2.23.0